+++ CHANGELOG.md
@@ -11,10 +11,12 @@ and this project adheres to [Semantic Versioning][semver].
 
 ### Changed
 
+- Use `_load_signing_keys` in `add_signing_key` ([138])
 - Raise a nicer error when instantiating a TUF repository if it is invalid ([137])
 
 ### Fixed
 
+[138]: https://github.com/openlawlibrary/taf/pull/138
 [137]: https://github.com/openlawlibrary/taf/pull/137
 
 ## [0.5.0] - 06/04/2020